+++ packages/opentelemetry-node/autoinstrumentation.js
@@ -1,0 +1,26 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch B.V. licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *	http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+// This file is included in the elastic-otel-node Docker image. It is used by
+// the OpenTeleemtry Operator for Kubernetes to support auto-instrumentation of
+// Node.js applications. The OTel Operator sets the following envvar for
+// Node.js apps to setup the OpenTelemetry SDK:
+//      --require .../autoinstrumentation.js
+
+require('./require.js');